more restructuring and generalizing  - commit to fix an issue
