+++ bot/trainer.py
